# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/booking": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Books a room for a user, requiring down payment based on the room price and user's balance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Book a room",
                "parameters": [
                    {
                        "description": "Booking Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.BookingRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Example: 'Bearer \u003cyour_token\u003e')",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "402": {
                        "description": "Insufficient balance",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Room or User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/booking/report": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all bookings made by a specific user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Get booking report for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token (Example: 'Bearer \u003cyour_token\u003e')",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking report retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "User not authorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/invoice_webhook_url": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Processes the webhook sent by Xendit for invoice updates (e.g., PAID, EXPIRED).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Handles Invoice Webhook from Xendit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Xendit Callback Token",
                        "name": "x-callback-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Webhook Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.WebhookPayload"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Example: 'Bearer \u003cyour_token\u003e')",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Failed to parse JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid callback token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/top-up": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Top up the balance of a user by providing a `BalanceRequest` object which includes the amount to be added to the balance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Saldo"
                ],
                "summary": "Top up user balance",
                "parameters": [
                    {
                        "description": "Top Up Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.BalanceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Example: 'Bearer \u003cyour_token\u003e')",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance successfully topped up",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid user ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.BalanceRequest": {
            "type": "object",
            "required": [
                "top_up_balance",
                "user_id"
            ],
            "properties": {
                "top_up_balance": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entity.BookingRequest": {
            "type": "object",
            "required": [
                "date_in",
                "date_out",
                "room_id"
            ],
            "properties": {
                "date_in": {
                    "type": "string"
                },
                "date_out": {
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.WebhookPayload": {
            "type": "object",
            "properties": {
                "adjusted_received_amount": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "bank_code": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "fees_paid_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "is_high": {
                    "type": "boolean"
                },
                "merchant_name": {
                    "type": "string"
                },
                "paid_amount": {
                    "type": "number"
                },
                "paid_at": {
                    "type": "string"
                },
                "payer_email": {
                    "type": "string"
                },
                "payment_channel": {
                    "type": "string"
                },
                "payment_destination": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}